+++ lib/src/clojure_lsp/handlers.clj
@@ -432,9 +432,10 @@
       (let [db @db*
             diagnostics (-> context :diagnostics)
             [row col] (shared/position->row-col (:start range))
+            [end-row end-col] (some-> (:end range) shared/position->row-col)
             root-zloc (parser/safe-zloc-of-file db (:uri text-document))
             client-capabilities (get db :client-capabilities)]
-        (f.code-actions/all root-zloc (:uri text-document) row col diagnostics client-capabilities db)))))
+        (f.code-actions/all root-zloc (:uri text-document) row col end-row end-col diagnostics client-capabilities db)))))
 
 (defn code-lens
   [{:keys [db*] :as components} {:keys [text-document]}]